% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotAllKmerCigars}
\alias{plotAllKmerCigars}
\title{Plot All Kmer CIGARs}
\usage{
plotAllKmerCigars(x, lg = F)
}
\arguments{
\item{x}{tidied kmer cigar data from tidyKmerCigar}
}
\value{
List of Plots
}
\description{
Plot all Sequences from tidied kmer cigar table, facet by seqid, and colour by taxa
}
