% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotKmerCigar}
\alias{plotKmerCigar}
\title{Plot k-mer CIGAR}
\usage{
plotKmerCigar(x, seqid)
}
\arguments{
\item{x}{tidied kmer cigar data from cleanKmerData}

\item{seqid}{seqence id}
}
\value{
Plot
}
\description{
Plot one sequence's CIGAR data
}
