% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot Summary Ridges}
\usage{
plotSummary(x)
}
\arguments{
\item{x}{tidied kmer cigar data from tidyKmerCigar}
}
\value{
Ridges plots
}
\description{
Create ridges plot that shows what proportion of kmers in all reads are assigned to each taxonomic group
}
