% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readKmerData}
\alias{readKmerData}
\title{Read in Kmer data}
\usage{
readKmerData(x)
}
\arguments{
\item{x}{path to kraken tsv}
}
\value{
tibble of kmer data with columns "Status", "SEQID", "TAXID", "Length", "CIGAR"
}
\description{
Returns a kmer tibble
}
