% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidyKmerData}
\alias{tidyKmerData}
\title{Tidy Kmer data in preparation to plot}
\usage{
tidyKmerData(x)
}
\arguments{
\item{x}{kraken tsv file imported with read.delim}
}
\value{
tibble of tidied kmer data: Status SEQID TAXID Length order taxid taxeng kmers CIGARPOS CIGARPOS2
}
\description{
Grabs the NCBI taxa names and returns a tidied kmer tibble
}
