% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidyPairedKmerData}
\alias{tidyPairedKmerData}
\title{Tidy Paired Kmer data in preparation to plot}
\usage{
tidyPairedKmerData(x)
}
\arguments{
\item{x}{kraken tsv file imported with read.delim}
}
\value{
tibble of tidied kmer data: Status SEQID TAXID Length order taxid taxeng kmers read CIGARstart CIGARstart
}
\description{
Grabs the NCBI taxa names and returns a tidied kmer tibble. Includes consideration for reads from paired data
}
